move packages, remove old config, fix logging
