+++ src/main/java/net/snackbag/tt20/TT20.java
@@ -3,9 +3,7 @@ package net.snackbag.tt20;
 import net.fabricmc.api.ModInitializer;
 
 import net.snackbag.tt20.command.CommandRegistry;
-import net.snackbag.tt20.config.BlockEntityMaskConfig;
-import net.snackbag.tt20.config.MainConfig;
-import net.snackbag.tt20.config.impl.TT20Config;
+import net.snackbag.tt20.config.TT20Config;
 import net.snackbag.tt20.util.TPSCalculator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,9 +15,7 @@ public class TT20 implements ModInitializer {
     public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 	public static final String VERSION = "0.7.1";
 	public static final TPSCalculator TPS_CALCULATOR = new TPSCalculator();
-	public static final MainConfig config = new MainConfig();
     public static final TT20Config Config = new TT20Config();
-	public static final BlockEntityMaskConfig blockEntityMaskConfig = new BlockEntityMaskConfig();
 
 	public static boolean warned = false;
