+++ src/main/java/net/snackbag/tt20/config/lib/types/CValue.java
@@ -1,4 +1,4 @@
-package net.snackbag.tt20.config.rework.types;
+package net.snackbag.tt20.config.lib.types;
 
 public abstract class CValue<T> {
     public final String key;
